Refactor ads complete step.
